Improve Builder.toMatrix
The previous implementation of builder was joining dynamic matrices
together vertically and creating a sized matrix. As horizontally
joining matrices is considerably cheaper, this is now done instead.

The builder now builds a list of columns, which are turned into a
list of rows, which is turned into the final matrix. Profiling has
shown this to be considerably quicker. The Builder module is still
an ungodly mess though.