Get dictionary for Tableau problems
Tableau problems can now output the current value of their objective
function and decision variables by calling toResult. Slack and
artificial variables are not shown, as there is no way to statically
specify all columns without knowing the phase. However, knowing the
phase would mean it is in the type of the solver, which would prevent
it moving from phase 1 to phase 2.