Introduce solveWith function to solve problems
Problems can now be solved using a solveWith functio in the toplevel
Solver module. This function takes a problem which can be solved by
the chosen solver, and the function to create the chosen solver. It
then solves the problem by repeatedly calling step until it reaches
optimality or fails.