Smart constructors for problem instances
To be able to tell when artificial variables are needed (without
being able to resort to dependent types), the construction of
problems has to become more restricted. The restrictions added
is that all RHS coefficients in constraints must be non-negative.
This means that all GEQ and EQU constraints need artificial variables,
and no LEQ constraints do.

Smart constructors to check these constraints are added to the
Problem module, and to make the module more uniform smart constructors
are added for the main Problem type as well. To still allow pattern
matching without exposing consructors, unidirectional pattern synonyms
are exposed from the module.