Add demo executable and MonadSolver class
A demo executable has been added to show how problems can be defined
and optimized using optimum. Currently the demo executable shows two
different problems, and has a main function which solves one of them
and prints the result.

There is now a class for monadic solvers, such as those using FFI.
This is implemented for all non-monadic solvers, using the Identity
monad as the base.