Phase I feasibility check; new Show instances
The Phase I part of the two phase simplex now checks whether the
artificial problem's solution indicates infeasiblity in the original
problem. The Show instances for the tableaus have been changed to
also print out the column and row variables.

The monadic class introduced in the last commit has been removed.
This is because monadic counterparts for other parts of the solver
interface would need to be made, and I don't want to commit to an
interface before considering implementing a monadic solver.